+++ lib/src/bones_ui_document.dart
@@ -1,5 +1,3 @@
-
-
 class URLLink {
   final String url;